keyword_idea_generator/requirements.txt
flask==2.3.3
requests==2.26.0
beautifulsoup4==4.10.0
playwright==1.16.0
pytrends==4.9.2
python-dotenv==0.19.2
sqlalchemy==1.4.27
